+++ test.js
@@ -25,7 +25,7 @@ test('Empty map', async (t) => {
     dpi: 300,
     layout: 'landscape_a4',
     format: 'pdf',
-    customAttributes: {},
+    customAttributes: {title: 'My title'},
     customizer: customizer,
   });
   assert.deepEqual(result, {
@@ -38,6 +38,7 @@ test('Empty map', async (t) => {
         rotation: 0,
         scale: 1,
       },
+      title: 'My title',
     },
     format: 'pdf',
     layout: 'landscape_a4',
@@ -45,7 +46,7 @@ test('Empty map', async (t) => {
 });
 
 test('OSM map', async (t) => {
-  const MFP_URL = 'https://geomapfish-demo-2-5.camptocamp.com/printproxy';
+  const MFP_URL = 'https://geomapfish-demo-2-8.camptocamp.com/printproxy';
   const layout = '1 A4 portrait'; // better take from MFP
   const map = new Map({
     target: 'map',